pyproject.toml
[project]
name = "alcove-search"
version = "0.2.0"
description = "Local-first document retrieval. Your data never leaves your disk."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "John Malone", email = "john.malone@gmail.com"},
]
keywords = ["retrieval", "search", "local-first", "chromadb", "documents"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Text Processing :: Indexing",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "chromadb>=0.5.5,<1.0",
    "fastapi>=0.115.0,<1.0",
    "uvicorn>=0.30.0,<1.0",
    "pypdf>=4.3,<5.0",
    "beautifulsoup4>=4.12,<5.0",
    "jinja2>=3.1,<4.0",
    "python-multipart>=0.0.6,<1.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.3", "python-docx>=1.0"]
docx = ["python-docx>=1.0"]
semantic = ["sentence-transformers>=3.0"]
epub = ["ebooklib>=0.18,<1.0"]
zvec = ["zvec>=0.1"]

[project.scripts]
alcove = "alcove.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["alcove"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.briefcase]
bundle = "com.alcove-search"
url = "https://github.com/Pro777/alcove"
